Update buzzer_app.html
I have tested your live URL and can see that when you click the "Create Game Session" button, it fails to connect to the database.

This is happening due to a subtle but critical JavaScript scoping issue. The createGame function is losing its connection to the main App object when the button is clicked, which prevents it from finding the initialized Supabase client.

To fix this, I will make the event handler more direct and robust, completely eliminating the scoping problem. I will also add more detailed error reporting in the alert message to make future debugging easier if needed.

Here is the update for buzzer_app.html: