Update buzzer_app.html
bug fix
-The problem is a subtle but critical JavaScript scoping issue. When the "Join Game" button is clicked, the function that runs is losing its connection to the main App.Client object, so it can't find the showBuzzerScreen function it's supposed to call.

To fix this definitively, I will restructure the code to use a more robust and standard way of handling events that guarantees the connection is never lost.

The Solution: Cleaner Event Handling
I am introducing a new function called handleRegistration within the Client object. We will then bind this function directly to the "Join Game" button's click event. This is a standard practice that completely eliminates the type of error you have been seeing.

I have also fixed a small related bug in the listenForBuzzes function on the Host view that was using an incorrect reference.